ui-warped-map: move stories to storybook
This way we can use proper imports from '@osrd-project/ui-warped-map'
instead of accessing directly internal files.
