+++ storybook/stories/WarpedMap/Algorithms.tsx
@@ -5,12 +5,8 @@ import type { Feature, LineString } from 'geojson';
 import { Layer, type LineLayer, Source } from 'react-map-gl/maplibre';
 
 import { OSM_BASE_MAP_STYLE, OSM_SOURCE } from './helpers';
-import BaseMap from '../components/BaseMap';
-import Loader from '../components/Loader';
-import WarpedMap from '../components/WarpedMap';
-import getWarping, { type WarpingOptions } from '../core/getWarping';
-import { type SourceDefinition } from '../core/types';
-import { useAsyncMemo } from '../core/useAsyncMemo';
+import { BaseMap, Loader, WarpedMap, getWarping, type WarpingOptions, type SourceDefinition } from '@osrd-project/ui-warped-map';
+import { useAsyncMemo } from './useAsyncMemo';
 
 const SOURCES: SourceDefinition[] = [OSM_SOURCE];
