+++ storybook/stories/WarpedMap/SampleMap.tsx
@@ -6,11 +6,8 @@ import 'maplibre-gl/dist/maplibre-gl.css';
 import { Layer, type LineLayer, Source } from 'react-map-gl/maplibre';
 
 import { OSM_BASE_MAP_STYLE, OSM_SOURCE } from './helpers';
-import BaseMap from '../components/BaseMap';
-import Loader from '../components/Loader';
-import WarpedMap from '../components/WarpedMap';
-import { type SourceDefinition } from '../core/types';
-import { useAsyncMemo } from '../core/useAsyncMemo';
+import { BaseMap, Loader, WarpedMap, type SourceDefinition } from '@osrd-project/ui-warped-map';
+import { useAsyncMemo } from './useAsyncMemo';
 
 const SOURCES: SourceDefinition[] = [OSM_SOURCE];
